requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.4
backports.zoneinfo==0.2.1
billiard==4.1.0
boto3==1.28.44
botocore==1.31.44
cachecontrol==0.13.1
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.3
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.5
django-amazon-ses==4.0.1
django-braces==1.15.0
django-celery-beat==2.5.0
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-filter==23.2
django-js-asset==2.1.0
django-oauth-toolkit==2.3.0
django-storages==1.14
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docutils==0.20.1
drf-tracking-logger==1.9.6
grpcio==1.58.0
gunicorn==21.2.0
h11==0.14.0
hiredis==2.2.3
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
jmespath==1.0.1
jwcrypto==1.5.0
kombu==5.3.2
more-itertools==10.1.0
msgpack==1.0.5
oauthlib==3.2.2
packaging==23.1
phonenumbers==8.13.20
Pillow==10.0.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.3
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pycurl==7.45.2
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stripe==6.4.0
toml==0.10.2
tomli==2.0.1
twisted==23.8.0
txaio==23.1.1
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.6.2
wrapt==1.15.0
zipp==3.16.2
zope.interface==6.0